notion-logger-agent/pyproject.toml
[tool.poetry]
name = "internet-of-agents"
version = "0.0.0"
description = "Internet of AI Agents"
authors = ["Multion"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
multion = "^1.3.3"
prefect = "^2.19.5"
instructor = "*"
openai = "*"
pydantic = "*"
graphviz = "*"
python-dotenv = "^1.0.1"
agentops = "0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
